Move `is_commuting` to `pennylane/ops/functions` (#2991)
* Move is_commuting

* Update doc/releases/changelog-dev.md

* add to introduction page
