+++ pennylane/ops/functions/__init__.py
@@ -20,14 +20,19 @@ This module contains functions that act on operators and tapes.
     :toctree: api
 
     ~eigvals
-    ~generator
-    ~matrix
     ~equal
+    ~generator
+    ~is_commuting
+    ~is_hermitian
+    ~is_unitary
+    ~matrix
+    ~simplify
 
 """
 from .eigvals import eigvals
 from .equal import equal
 from .generator import generator
+from .is_commuting import is_commuting
 from .is_hermitian import is_hermitian
 from .is_unitary import is_unitary
 from .matrix import matrix
